Official Patch 1
If this works (I've somewhat tested it) it should automagically patch the surface API that we're given by cobalt. I've had it fixed clientside for awhile, but I forgot that it wasn't a global fix.

I also added worldmap.nfp because I forgot to add it in the last alpha. Woops.

(Expect Official Patch 2 if this one still turns out to be flawed)