Official Patch 2
This has nothing to do with official patch 1, so that's why it's named official patch 2... Don't judge my naming conventions.

But seriously, this is fixing the ability to build the strategic command center on the view button, as well as preventing the ability to switch views when the game is waiting for you to build the SCC.

Not a big deal, just didn't want to wait until the next alpha to fix it.