Impending Doom Alpha 4
Hurray! The kind of playable update! Major changes include:

The ability to fire nukes at enemy cities

The beginning of the ingame help system (doomapedia)

Config file and working options menu

Cleaner installation of Cobalt (no longer a shell.run command)

Semi-automatic updating available within the game, as well as a notification if you're not on the latest version

Paved the way for major updates in Alpha 5

Several minor changes and bug fixes